move behaviors to behaviors directory
