+++ test/token/ERC20/ERC20Burnable.test.js
@@ -1,4 +1,4 @@
-const { shouldBehaveLikeERC20Burnable } = require('./ERC20Burnable.behavior');
+const { shouldBehaveLikeERC20Burnable } = require('./behaviors/ERC20Burnable.behavior');
 const ERC20BurnableMock = artifacts.require('ERC20BurnableMock');
 
 contract('ERC20Burnable', function ([_, owner, ...otherAccounts]) {
