+++ test/token/ERC20/behaviors/ERC20Burnable.behavior.js
