+++ test/token/ERC20/behaviors/ERC20Capped.behavior.js
