+++ test/token/ERC20/behaviors/RBACMintableToken.behavior.js
